Analyze performance between the human and AI screening.

When both the human and AI title and abstract screening has been done, this function allows you to calculate performance measures of the screening, including the overall accuracy, specificity and sensitivity of the screening, as well as interrater reliability kappa statistics.

Usage

screen_analyzer(x, human_decision = human_code)

Arguments

x

Either an object of class 'chatgpt' or a data set of class 'chatgpt_tbl'

human_decision

Indicate the variable in the data that contains the human_decision. This variable must be numeric containing 1 (for included references) and 0 (for excluded references) only.

Value

A tibble with screening performance measures. The tibble includes the following variables:

promptid	integer	indicating the prompt ID.

model	character	indicating the specific gpt-model used.

reps	integer	indicating the number of times the same question was sent to ChatGPT.

top_p	numeric	indicating the applied top_p.

n_screened	integer	indicating the number of screened references.

n_missing	numeric	indicating the number of missing responses.

n_refs	integer	indicating the number of references expected to be screened for the given condition.

human_in_gpt_ex	numeric	indicating the number of references included by humans and excluded by gpt.

human_ex_gpt_in	numeric	indicating the number of references excluded by humans and included by gpt.

human_in_gpt_in	numeric	indicating the number of references included by humans and included by gpt.

human_ex_gpt_ex	numeric	indicating the number of references excluded by humans and excluded by gpt.

accuracy	numeric	indicating the overall percent disagreement between human and gpt (Gartlehner et al., 2019).

p_agreement	numeric	indicating the overall percent agreement between human and gpt.

precision	numeric	"measures the ability to include only articles that should be included" (Syriani et al., 2023).

recall	numeric	"measures the ability to include all articles that should be included" (Syriani et al., 2023).

npv	numeric	Negative predictive value (NPV) "measures the ability to exclude only articles that should be excluded" (Syriani et al., 2023).

specificity	numeric	"measures the ability to exclude all articles that should be excluded" (Syriani et al., 2023).

bacc	numeric	"capture the accuracy of deciding both inclusion and exclusion classes" (Syriani et al., 2023).

F2	numeric	F-measure that "consider the cost of getting false negatives twice as costly as getting false positives" (Syriani et al., 2023).

mcc	numeric	indicating percent agreement for excluded references (Gartlehner et al., 2019).

irr	numeric	indicating the interrater reliability as described in McHugh (2012).

se_irr	numeric	indicating standard error for the interrater reliability.

cl_irr	numeric	indicating lower confidence interval for the interrater reliability.

cu_irr	numeric	indicating upper confidence interval for the interrater reliability.

level_of_agreement	character	interpretation of the interrater reliability as suggested by McHugh (2012).

References

Gartlehner, G., Wagner, G., Lux, L., Affengruber, L., Dobrescu, A., Kaminski-Hartenthaler, A., & Viswanathan, M. (2019). Assessing the accuracy of machine-assisted abstract screening with DistillerAI: a user study. Systematic Reviews, 8(1), 277. https://doi.org/10.1186/s13643-019-1221-3

McHugh, M. L. (2012). Interrater reliability: The kappa statistic. Biochemia Medica, 22(3), 276-282. https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3900052/

Syriani, E., David, I., & Kumar, G. (2023). Assessing the Ability of ChatGPT to Screen Articles for Systematic Reviews. ArXiv Preprint ArXiv:2307.06464.

Examples

x <- AIscreenR:::result_object
x |> screen_analyzer() |> print(width=260)
#> # A tibble: 12 × 25
#>    promptid model               reps top_p n_refs n_screened n_missing
#>       <int> <chr>              <int> <dbl>  <int>      <int>     <int>
#>  1        1 gpt-3.5-turbo-0613     1 0.001    135        132         3
#>  2        1 gpt-3.5-turbo-0613    10 0.001    135        135         0
#>  3        1 gpt-3.5-turbo-0613     1 1        135        135         0
#>  4        1 gpt-3.5-turbo-0613    10 1        135        135         0
#>  5        1 gpt-4                  1 0.001    135        135         0
#>  6        1 gpt-4                  1 1        135        134         1
#>  7        2 gpt-3.5-turbo-0613     1 0.001    135        135         0
#>  8        2 gpt-3.5-turbo-0613    10 0.001    135        135         0
#>  9        2 gpt-3.5-turbo-0613     1 1        135        135         0
#> 10        2 gpt-3.5-turbo-0613    10 1        135        135         0
#> 11        2 gpt-4                  1 0.001    135        135         0
#> 12        2 gpt-4                  1 1        135        133         2
#>    human_in_gpt_ex human_ex_gpt_in human_in_gpt_in human_ex_gpt_ex accuracy
#>              <dbl>           <dbl>           <dbl>           <dbl>    <dbl>
#>  1               8               7              36              81   0.111 
#>  2               8               8              37              82   0.119 
#>  3              11               8              34              82   0.141 
#>  4               4               8              41              82   0.0889
#>  5               1               9              44              81   0.0741
#>  6               0               6              45              83   0.0444
#>  7               9               6              36              84   0.111 
#>  8               8               7              37              83   0.111 
#>  9              10               6              35              84   0.119 
#> 10               7               7              38              83   0.104 
#> 11               1               8              44              82   0.0667
#> 12               2               7              43              81   0.0667
#>    p_agreement precision recall   npv specificity  bacc    F2   mcc   irr
#>          <dbl>     <dbl>  <dbl> <dbl>       <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1       0.889     0.837  0.818 0.910       0.920 0.869 0.517 0.871 0.762
#>  2       0.881     0.822  0.822 0.911       0.911 0.867 0.514 0.867 0.733
#>  3       0.859     0.810  0.756 0.882       0.911 0.833 0.489 0.839 0.678
#>  4       0.911     0.837  0.911 0.953       0.911 0.911 0.545 0.903 0.804
#>  5       0.926     0.830  0.978 0.988       0.9   0.939 0.561 0.924 0.840
#>  6       0.956     0.882  1     1           0.933 0.966 0.586 0.954 0.905
#>  7       0.889     0.857  0.8   0.903       0.933 0.867 0.517 0.873 0.746
#>  8       0.889     0.841  0.822 0.912       0.922 0.872 0.520 0.874 0.749
#>  9       0.881     0.854  0.778 0.894       0.933 0.856 0.509 0.864 0.727
#> 10       0.896     0.844  0.844 0.922       0.922 0.883 0.528 0.883 0.767
#> 11       0.933     0.846  0.978 0.988       0.911 0.944 0.567 0.931 0.856
#> 12       0.933     0.86   0.956 0.976       0.920 0.938 0.566 0.928 0.860
#>     se_irr cl_irr cu_irr level_of_agreement
#>      <dbl>  <dbl>  <dbl> <chr>             
#>  1 0.0126   0.737  0.787 Moderate          
#>  2 0.0124   0.709  0.758 Moderate          
#>  3 0.0131   0.652  0.704 Moderate          
#>  4 0.0111   0.783  0.826 Strong            
#>  5 0.0105   0.820  0.861 Strong            
#>  6 0.00831  0.889  0.921 Almost perfect    
#>  7 0.0118   0.723  0.769 Moderate          
#>  8 0.0120   0.725  0.772 Moderate          
#>  9 0.0121   0.704  0.751 Moderate          
#> 10 0.0117   0.744  0.790 Moderate          
#> 11 0.00991  0.836  0.875 Strong            
#> 12 0.0102   0.840  0.880 Strong