Approximate price estimation for title and abstract screening using ChatGPT models

This function supports the approximation of the price of the total title and abstract screening with ChatGPT. The function only provide approximately accurate pricing estimates for screenings drawing on the function calling function incl_function_simple (Default) since the detailed descriptions will increase the completion tokens with an unknown amount which furthermore is rather model sensitive with gpt-4 models yielding longer detailed responses relative to gpt-3.5-turbo models.

Usage

approximate_price_gpt(
  data,
  prompt,
  studyid,
  title,
  abstract,
  model = "gpt-3.5-turbo-0613",
  reps = 1,
  top_p = 1,
  token_word_ratio = 1.6
)

Arguments

data

Dataset containing the titles and abstracts.

prompt

Prompt(s) to paste before the title and abstract.

studyid

Unique ID of study. If missing, this is generated automatically.

title

Name of the variable containing the title information.

abstract

Name of variable containing the abstract information.

model

Character string with the name of the completion model. Can take multiple models, including gpt-4 models. Default = "gpt-3.5-turbo-0613". Find available model at https://platform.openai.com/docs/models/model-endpoint-compatibility.

reps

Numerical value indicating the number of times the same question should be sent to ChatGPT. This can be useful to test consistency between answers. Default is 1 but when using 3.5 models, we recommend setting this value to 10.

top_p

'An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered. We generally recommend altering this or temperature but not both.' (OPEN-AI). Default is 1. Find documentation at https://platform.openai.com/docs/api-reference/chat/create#chat/create-top_p.

token_word_ratio

The multiplier used to approximate the number of tokens per word. Default is 1.6 which we empirically have found to be the average number of tokens per word.

Value

An object of class "gpt_price". The object is a list containing the following components:

price

numerical value indicating the total approximate price (in USD) of the screening across all gpt-models expected to be used for the screening.

price_data

dataset with prices across all gpt models expected to be used for screening.

Examples

prompt <- "This is a prompt"

app_price <- approximate_price_gpt(
  data = filges2015_dat[1:2,],
  prompt = prompt,
  studyid = studyid,
  title = title,
  abstract = abstract,
  model = c("gpt-3.5-turbo-0613", "gpt-4"),
  reps = c(10, 1)
)

app_price
#> The approximate price of the (simple) screening will be around $0.0674.
app_price$price_dollar
#> [1] 0.0674
app_price$price_data
#> # A tibble: 2 × 5
#>   model     iterations input_price_dollar output_price_dollar total_price_dollor
#>   <chr>          <dbl>              <dbl>               <dbl>              <dbl>
#> 1 gpt-3.5-…         10             0.0219             0.00044             0.0223
#> 2 gpt-4              1             0.0438             0.00132             0.0451