Analyze performance between the human and AI screening.

When both the human and AI title and abstract screening has been done, this function allows you to calculate performance measures of the screening, including the overall accuracy, specificity and sensitivity of the screening, as well as interrater reliability kappa statistics.

Usage

screen_analyzer(x, human_decision = human_code)

Arguments

x

Either an object of class 'chatgpt' or a data set of class 'chatgpt_tbl'

human_decision

Indicate the variable in the data that contains the human_decision. This variable must be numeric containing 1 (for included references) and 0 (for excluded references) only.

Value

A tibble with screening performance measures. The tibble includes the following variables:

promptid	integer	indicating the prompt ID.

model	character	indicating the specific gpt-model used.

reps	integer	indicating the number of times the same question was sent to ChatGPT.

top_p	numeric	indicating the applied top_p.

n_screened	integer	indicating the number of screened references.

n_missing	numeric	indicating the number of missing responses.

n_refs	integer	indicating the number of references expected to be screened for the given condition.

human_in_gpt_ex	numeric	indicating the number of references included by humans and excluded by gpt.

human_ex_gpt_in	numeric	indicating the number of references excluded by humans and included by gpt.

human_in_gpt_in	numeric	indicating the number of references included by humans and included by gpt.

human_ex_gpt_ex	numeric	indicating the number of references excluded by humans and excluded by gpt.

accuracy	numeric	indicating the overall percent disagreement between human and gpt (Gartlehner et al., 2019).

p_agreement	numeric	indicating the overall percent agreement between human and gpt.

sensitivity	numeric	indicating percent agreement for included references (Gartlehner et al., 2019).

specificity	numeric	indicating percent agreement for excluded references (Gartlehner et al., 2019).

IRR	numeric	indicating the interrater reliability as described in McHugh (2012).

SE_IRR	numeric	indicating standard error for the interrater reliability.

CL_IRR	numeric	indicating lower confidence interval for the interrater reliability.

CU_IRR	numeric	indicating upper confidence interval for the interrater reliability.

level_of_agreement	character	interpretation of the interrater reliability as suggested by McHugh (2012).

References

Gartlehner, G., Wagner, G., Lux, L., Affengruber, L., Dobrescu, A., Kaminski-Hartenthaler, A., & Viswanathan, M. (2019). Assessing the accuracy of machine-assisted abstract screening with DistillerAI: a user study. Systematic Reviews, 8(1), 277. https://doi.org/10.1186/s13643-019-1221-3

McHugh, M. L. (2012). Interrater reliability: The kappa statistic. Biochemia Medica, 22(3), 276-282. https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3900052/

Examples

x <- AIscreenR:::result_object
x |> screen_analyzer() |> print(width=220)
#> # A tibble: 6 × 20
#>   promptid model          reps top_p n_screened n_missing n_refs human_in_gpt_ex
#>      <int> <chr>         <int> <dbl>      <int>     <dbl>  <int>           <dbl>
#> 1        1 gpt-3.5-turb…     1     1        135         0    135              11
#> 2        1 gpt-3.5-turb…    10     1        135         1    135               4
#> 3        1 gpt-4             1     1        134         1    135               0
#> 4        2 gpt-3.5-turb…     1     1        135         0    135              10
#> 5        2 gpt-3.5-turb…    10     1        135         0    135               7
#> 6        2 gpt-4             1     1        133         2    135               2
#>   human_ex_gpt_in human_in_gpt_in human_ex_gpt_ex accuracy p_agreement
#>             <dbl>           <dbl>           <dbl>    <dbl>       <dbl>
#> 1               8              34              82   0.141        0.859
#> 2               8              41              82   0.0889       0.911
#> 3               6              45              83   0.0444       0.956
#> 4               6              35              84   0.119        0.881
#> 5               7              38              83   0.104        0.896
#> 6               7              43              81   0.0667       0.933
#>   sensitivity specificity   IRR  SE_IRR CL_IRR CU_IRR level_of_agreement
#>         <dbl>       <dbl> <dbl>   <dbl>  <dbl>  <dbl> <chr>             
#> 1       0.756       0.911 0.678 0.0131   0.652  0.704 Moderate          
#> 2       0.911       0.911 0.804 0.0111   0.783  0.826 Strong            
#> 3       1           0.933 0.905 0.00831  0.889  0.921 Almost perfect    
#> 4       0.778       0.933 0.727 0.0121   0.704  0.751 Moderate          
#> 5       0.844       0.922 0.767 0.0117   0.744  0.790 Moderate          
#> 6       0.956       0.920 0.860 0.0102   0.840  0.880 Strong