Title and abstract screening with ChatGPT using function calls.

This function supports the conduct of title and abstract screening with ChatGPT in R. The function allows to run title and abstract screening across multiple prompts and with repeated questions to check for consistency across answers. This function draws on the newly developed function calling to better steer output of response.

Usage

tabscreen_gpt(
  data,
  prompt,
  studyid,
  title,
  abstract,
  ...,
  arrange_var = studyid,
  model = "gpt-3.5-turbo-0613",
  role = "user",
  functions = incl_function_simple,
  function_call_name = list(name = "inclusion_decision_simple"),
  top_p = 1,
  time_info = TRUE,
  token_info = TRUE,
  api_key = get_api_key(),
  max_tries = 4,
  max_seconds = NULL,
  is_transient = gpt_is_transient,
  backoff = NULL,
  after = NULL,
  rpm = 3500,
  reps = 1,
  seed = NULL,
  progress = TRUE
)

Arguments

data

Data with title and abstracts

prompt

Prompt(s) to paste before title and abstract

studyid

ID indication unique ID of study. Is missing, this is generated automatically.

title

Name of variable with containing title information

abstract

Name of variable with containing abstract information

...

Further argument to pass to the request body. See https://platform.openai.com/docs/api-reference/chat/create.

arrange_var

Function indicating the variables determining the arrangement of the data. Default is studyid.

model

Character string with the name of the completion model. Can take multiple models, including gpt-4 models.Default = "gpt-3.5-turbo-0613". Find available model at https://platform.openai.com/docs/models/model-endpoint-compatibility.

role

Character string indicate the role of the user. Default is "user".

functions

Function to steer output. Default is incl_function_simple. Find further documentation for function calling at https://openai.com/blog/function-calling-and-other-api-updates.

function_call_name

Functions to call. Default is list(name = "inclusion_decision_simple")

top_p

'An alternative to sampling with temperature, called nucleus sampling, where the model considers the results of the tokens with top_p probability mass. So 0.1 means only the tokens comprising the top 10% probability mass are considered. We generally recommend altering this or temperature but not both.' (OPEN-AI). Find documentation at https://platform.openai.com/docs/api-reference/chat/create#chat/create-top_p.

time_info

Logical indicating whether the run time of each request/question should be included in the data. Default = TRUE.

token_info

Logical indicating whether the total number of tokens per request should be included in the output data. Default = TRUE.

api_key

Numerical value with your personal API key. Find at https://platform.openai.com/account/api-keys. Use secret_make_key(), secret_encrypt(), and secret_decrypt() to scramble and decrypt the api key and use set_api_key() to securely automate the use of the api key by setting the api key as a locale environment variable.

max_tries, max_seconds

'Cap the maximum number of attempts with max_tries or the total elapsed time from the first request with max_seconds. If neither option is supplied (the default), req_perform() will not retry' (Wickham, 2023).

is_transient

'A predicate function that takes a single argument (the response) and returns TRUE or FALSE specifying whether or not the response represents a transient error' (Wickham, 2023).

backoff

'A function that takes a single argument (the number of failed attempts so far) and returns the number of seconds to wait' (Wickham, 2023).

after

'A function that takes a single argument (the response) and returns either a number of seconds to wait or NULL, which indicates that a precise wait time is not available that the backoff strategy should be used instead' (Wickham, 2023).

rpm

Numerical value indicating the number of requests per minute (rpm) available for the specified api key. Find more information at https://platform.openai.com/docs/guides/rate-limits/what-are-the-rate-limits-for-our-api. Alternatively, use rate_limits_per_minute().

reps

Numerical value indicating the number of times the same question should be sent to ChatGPT. This can be useful to test consistency between answers. Default is 1.

seed

Numerical value for a seed to ensure that proper, parallel-safe random numbers are produced.

progress

Logical indicating whether a progress line should be shown when running the title and abstract screening in parallel. Default is TRUE.

Value

A tibble with the gpt decision, run time, tokens used, top_p, and number of repeated requests.

References

Wickham H (2023). httr2: Perform HTTP Requests and Process the Responses. https://httr2.r-lib.org, https://github.com/r-lib/httr2.

Examples

if (FALSE) {
tabscreen_gpt(
  data = FFT_dat[1:2,],
  prompt = prompt,
  studyid = studyid,
  title = title,
  abstract = abstract,
  model = c("gpt-3.5-turbo-0613", "gpt-3.5-turbo-16k-0613"),
  max_tries = 1,
  reps = 10
  )
 }